+++ src/components/form/FormFieldsRenderer.tsx
@@ -121,7 +121,7 @@ function renderField<
         case 'object':
         case 'tags':
         case undefined: {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <InputField
                     type="text"
@@ -135,7 +135,7 @@ function renderField<
             );
         }
         case 'number': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <InputField
                     type="number"
@@ -146,29 +146,23 @@ function renderField<
                 />
             );
         }
-        case 'integer': {
-            const { type, transform, validation, ...props } = fieldDefinition;
-            return (
-                <InputField
-                    type="number"
-                    pattern="\d*"
-                    {...props}
-                    showLabels={showLabels}
-                    onChange={e => onChange(transform?.output(e) ?? e)}
-                    {...(transform?.input(value) ?? { value })}
-                    {...field}
-                    errors={fieldState.invalid && fieldState.error ? [fieldState.error] : undefined}
-                />
-            );
-        }
+        case 'integer':
         case 'decimal': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <InputField
                     type="number"
+                    pattern={type === 'integer' ? '\\d*' : undefined}
                     {...props}
                     showLabels={showLabels}
-                    onChange={e => onChange(transform?.output(e) ?? e)}
+                    onChange={e =>
+                        onChange(
+                            transform?.output(e) ??
+                                (Number.isNaN(e.target.valueAsNumber)
+                                    ? undefined
+                                    : e.target.valueAsNumber)
+                        )
+                    }
                     {...(transform?.input(value) ?? { value })}
                     {...field}
                     errors={fieldState.invalid && fieldState.error ? [fieldState.error] : undefined}
@@ -176,7 +170,7 @@ function renderField<
             );
         }
         case 'email': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <InputField
                     type="email"
@@ -190,7 +184,8 @@ function renderField<
             );
         }
         case 'password': {
-            const { type, transform, validation, withPolicyRules, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, withPolicyRules, ...props } =
+                fieldDefinition;
             return (
                 <PasswordField
                     {...props}
@@ -205,7 +200,7 @@ function renderField<
             );
         }
         case 'date': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <DateField
                     {...props}
@@ -218,7 +213,7 @@ function renderField<
             );
         }
         case 'checkbox': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <CheckboxField
                     {...props}
@@ -231,7 +226,7 @@ function renderField<
             );
         }
         case 'radio-group': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <RadioGroupField
                     {...props}
@@ -244,7 +239,7 @@ function renderField<
             );
         }
         case 'select': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <SelectField
                     {...props}
@@ -257,7 +252,8 @@ function renderField<
             );
         }
         case 'phone': {
-            const { type, transform, validation, phoneNumberOptions, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, phoneNumberOptions, ...props } =
+                fieldDefinition;
             return (
                 <PhoneNumberField
                     {...props}
@@ -277,7 +273,7 @@ function renderField<
             );
         }
         case 'hidden': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <input
                     type="hidden"
@@ -289,7 +285,7 @@ function renderField<
             );
         }
         case 'identifier': {
-            const { type, transform, validation, ...props } = fieldDefinition;
+            const { type, transform, validation, defaultValue, ...props } = fieldDefinition;
             return (
                 <IdentifierField
                     {...props}